"ads2_m8_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2-assignments/m8/1. Shortest Paths/Solution.java
@@ -47,7 +47,7 @@ public final class Solution {
 			int vertFrom  = stations.get(inp[0]);
 			int vertTo  = stations.get(inp[1]);
 			DijkstraUndirectedSP dusp = new DijkstraUndirectedSP(eg, vertFrom);
-			System.out.println(dusp.distTo(vertTo));
+			System.out.println((int) dusp.distTo(vertTo));
 		}
 	}
 }
